+++ app/(route)/activity/(form)/_components/BannerImage/index.tsx
