+++ app/(route)/activity/(form)/_components/IntroduceImage/index.tsx
