+++ app/(route)/activity/(form)/register/page.tsx
